move solar adjustment into a new file
